share sidekick logic
