+++ src/components/Sidekick/copilot.tsx
