+++ src/screens/surrealist/views/sidekick/SidekickView/index.tsx
@@ -1,14 +1,7 @@
-import { useAbsoluteLocation, useViewFocus } from "~/hooks/routing";
+import { Sidekick } from "~/components/Sidekick";
 
 export function SidekickView() {
-	const [, navigate] = useAbsoluteLocation();
-
-	// For now, redirect to the chat page when the sidekick is focused
-	useViewFocus("sidekick", () => {
-		navigate("/chat");
-	});
-
-	return null;
+	return <Sidekick />;
 }
 
 export default SidekickView;
